Output FASTQC to trim folder without align warning
This commit re-introduces the fastqc reports to the output from the trimming process so they are output to the trim folder.

This commit also outputs the original path to the forward and reverse reads from the trimming process so the align process can correctly parse the path to the input fastqc reports. The align process does not use them, but must correctly accept them for the trimming process to output them to the trim folder.